requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
channels>=4.0
channels-redis>=4.2
daphne>=4.1
celery>=5.4
redis>=5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
django-environ>=0.11
cryptography>=42.0
httpx>=0.27
groq>=0.11
anthropic>=0.34
ollama>=0.3
tiktoken>=0.7
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
